rename js files
