+++ package.json
@@ -3,7 +3,14 @@
   "version": "1.0.0",
   "description": "VT Code Camp website",
   "main": "index.js",
+  "scripts": {
+    "build": "npx eleventy",
+    "serve": "npx eleventy --serve",
+    "update-data": "node tools/update-data.js",
+    "update-and-build": "npm run update-data && npm run build"
+  },
   "dependencies": {
+    "@11ty/eleventy": "^1.0.1",
     "clean-css": "^5.3.1",
     "eleventy-plugin-nesting-toc": "^1.3.0",
     "got": "^11.8.5",
@@ -11,15 +18,6 @@
     "moment": "^2.29.4",
     "node-fetch": "2.6",
     "sharp": "^0.30.7"
-  },
-  "devDependencies": {
-    "@11ty/eleventy": "^1.0.1"
-  },
-  "scripts": {
-    "build": "npx eleventy",
-    "serve": "npx eleventy --serve",
-    "update-data": "node tools/updateData.js",
-    "update-and-build": "npm run update-data && npm run build"
   },
   "repository": {
     "type": "git",
