+++ tools/update-photos.js
